Start to work on configurable Dashboard view, extract some reusable components.
