+++ client/js/src/main/scala/zio/zmx/client/frontend/components/LayoutConstants.scala
@@ -1,0 +1,5 @@
+package zio.zmx.client.frontend.components
+
+object LayoutConstants {
+  lazy val layoutPanel: String = "p-2 bg-neutral text-neutral-content rounded-box text-lg font-bold"
+}
