+++ client/js/src/main/scala/zio/zmx/client/frontend/model/DashboardConfig.scala
@@ -1,0 +1,5 @@
+package zio.zmx.client.frontend.model
+
+final case class DashboardConfig(
+  name: String
+)
