+++ client/js/src/main/scala/zio/zmx/client/frontend/views/DashboardView.scala
@@ -1,0 +1,5 @@
+package zio.zmx.client.frontend.views
+
+object DashboardView {
+  
+}
